+++ R/pat_createNew.R
@@ -24,6 +24,7 @@
 #' 
 #' @examples
 #' \donttest{
+#' setArchiveBaseUrl("http://smoke.mazamascience.com/data/PurpleAir")
 #' pas <- pas_load()
 #' pat <- pat_createNew(pas, "Seattle", startdate = 20180701, enddate = 20180901)
 #' pat_multiplot(pat)