+++ R/pat_externalFit.R
@@ -36,6 +36,7 @@
 #' 
 #' @examples
 #' \donttest{
+#' setArchiveBaseUrl("http://smoke.mazamascience.com/data/PurpleAir")
 #' pat <- pat_load("SCAH_22")
 #' pat_externalFit(pat)
 #' }